multi-model-comfyui/requirements.txt
runpod>=1.5.1
requests>=2.31.0
boto3>=1.26.137
Pillow>=10.0.0
opencv-python>=4.8.0.74
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
xformers>=0.0.20
compel>=2.0.0
controlnet-aux>=0.0.7